# Conflicts:
#	lib/abstract_game/GameIDs.hpp
#	lib/scene/DashboardScene.cpp
#	src/App.cpp

+++ lib/abstract_game/GameRunThrough.cpp
@@ -8,7 +8,7 @@ namespace abstract_game {
 
 
     GameRunThrough::GameRunThrough(size_t const &gameSessionUID, int const &gameRunThroughUID,
-                                   std::string const &resultUnit, long const &result)
+                                   std::string const &resultUnit, double const &result)
             : gameSessionUID{gameSessionUID}, gameRunThroughUID{gameRunThroughUID}, resultUnit{resultUnit},
               result{result} {
